Add read-only support for the SpaceChem leaderboard
Adds all the models for SpaceChem, plus an implementation
of a scraper for the reddit wiki.
There is no update implemented, because the process needs to remain manual.

The implementation is in java (because I don't know Kotlin), the build
system has been modified slightly to add the needed dependencies